+++ lib/engine_graphics/src/bitmap/loaders/vsx_bitmap_loader_jpg.cpp
@@ -10,7 +10,7 @@ void vsx_bitmap_loader_jpg::worker(vsx_bitmap* bitmap, vsx::filesystem* filesyst
   njInit();
 
   unsigned char* data = (unsigned char*)filesystem->f_gets_entire(file);
-  nj_result_t result = njDecode(data, filesystem->f_get_size(file));
+  nj_result_t result = njDecode(data, (int)filesystem->f_get_size(file));
   filesystem->f_close(file);
 
   req(result == NJ_OK);
